requirements.txt
# DesignBridge dependencies
langgraph>=1.0.0
langchain-core>=0.1.0
typing-extensions>=4.0.0
streamlit>=1.30.0

# Gemini API
google-generativeai>=0.3.0

# Local vision inference (Depth + Segmentation)
numpy>=1.24.0
pillow>=10.0.0
transformers>=4.41.0
torch>=2.2.0
torchvision>=0.17.0
